Remove groups from specliast browse pages
The groups weren't testing very well and were deemed no longer needed.
Remove the groups from the frontend and update the tests.

Introduced a presenter for the specialist sectors which automatically
remove the sector from the artefact title. This made sorting and
displaying by title easier to follow. Removed the helper that previously
held this functionality.

Updated tests to work with just a list of artefacts rather than a nested
by group.