Style the parent tag title on subsector pages
As we no longer use the full wrapper template for collections, we’ve 
lost the styling on subsector pages for the title of the parent sector.

As we don’t use this style anywhere near as frequently as we used to on
GOV.UK, it feels like this style should be an application concern, so 
here I’ve copied the rules from static into the specialist sectors 
stylesheet.