Allow popState to load new data
If you are using the back button to the page which was rendered on the
server we don't get the ajax cache passed in the state object. This lets
us reload that data from the server if the data isn't in the state.

This also moves the title logic from getting the title from the clicked
link to finding the link on the page which targets the current page.
This means that we can reload the title on back button.