Move focus to headings and remove aria-live
Having the focus (and relevent tabindex) sent to the column wrappers
caused browsers to jump down to try and get as much of the column in
view as possible. This was visually distracting. Moving the focus (and
tabindex) to the heading at the top of the column. This removed the
visuall jump.

Used deferreds to send the focus to the heading after the animation has
completed.

In trying to fix this I also challenged my assumptions about how the
screen readers were reading it and listened. Having the entire block
wrapped in an aria-live forced the screen reader to read out all of the
new content. This is very unhelpful. As we are already now having the
correct heading focused and thus read out reading out anything extra is
superfluous. So I removed the aria-live.