Adds services registration/initialization module
* Moves services registration/initialization out of ApplicationController
* Changes code to use Collections#services

Having the services initialized with the app caused changes to
BrowseControllerTest. This test previously stubbed Plek#website_root in
a before block, to ensure a consistent root for the URLs.

This PR maintains that behaviour by setting the GOVUK_WEBSITE_ROOT env var in the test helper, before
setting the Rails env.