Pass 'start' and 'count' parameters to Subcategory
This changes the SubcategoriesController to pass the 'start' and 'count'
pagination parameters through to the Subcategory model when fetching a
sub-topic.

This adds some basic sanitisation of the parameters, forcing them to be
integers, and removes any blank parameters so that we can maintain
defaults further upstream.

This also improves the test coverage around the latest_changes action of
the controller.