Display pagination on the latest changes view
This changes the latest changes view template to display previous and
next page links. The view uses the ChangedDocumentsPaginationPresenter
to infer whether a next or previous page link should be displayed.

The markup is identical to the pagination links used for search results
in Frontend, and I've copied the stylesheet and arrow sprite image over
from there too.

This also includes a full integration test for the pagination, which
stubs the API requests using the Collections API adapter's test helper.