Remove beta header from subtopic show pages.
To quote the ticket:
Now that sub-topics have curated groups, email alerts and a latest feed, and now that they'll be the official (and only) source of content about topics for many different government organisations, we should no longer be denoting them as "beta".

We should still show the "beta" label on topic pages, as we know we still want to iterate these significantly, and the latest changes feed, as not all content have timestamps and change notes.