Use a custom beta message on the latest changes feed
Whilst we want to continue showing a beta label on the latest changes
feed, the beta label provided by the Slimmer processor (which says that
the page is 'being rebuilt') isn't accurate.

This replaces the use of the Slimmer beta notice processor with the use
of the govuk_component shared beta notice template, which allows us to
insert a custom message.