Pay attention to the SCHEMA_GIT_COMMIT variable
If this is set, use this in calls to gh-status, and set the
govuk-content-schemas repo to be at this commit.

If it's not set, use the "GIT_COMMIT" setting for status messages, since
we'll be sending messages for our repo, and just use master for
govuk-content-schemas ("git checkout" with no parameter will do
nothing).