Replace external Ajax calls with internal partial rendering
When browsing the browse pages, the JS currently requests the data from
the content-store directly and renders a Mustache partial with that
data. Serverside the exact same thing happens.

This commit rewrites the JS so that instead of requesting and rendering
raw data, it requests the partial directly from the Rails app.

More cleanup in future commits.