Remove unnecessary helpers for integration tests
In some cases helpers are being over-used in the integration tests.
Causing the reader to look up a step definition, and then a helper, to
understand what a cucumber step is doing. This commit removes the
indirection of examples where helpers are only used once, or where the
actual implementation is easier to understand than the helper.