Fix jenkins.sh error trap.
The `trap` line needs to be single-quoted so that `$LINENO` is evaluated
at the point of the error, not the point the trap is defined.

Also fix the status reported on error to be error as opposed to failure.
This will make it easier to determine the nature of the failure.