Support both namespaced and non-namespaced topic paths.
While topics are being migrated from the root namespace to the /topic
namespace, this needs to support both.  Once the migration is complete,
support for the root namespace can be removed.

This creates some temporary helper methods that wrap the corresponging
route helpers, selecting the appropriate one based on the request path.

This also updates the integration tests to use the namespaced paths.
The cucumber tests remain using the non-namespaced paths to give some
coverage of both variants.