Remove routes for topics without namespace
Topics have been moved to the /topic namespace. This commit removes all
transitional code to serve topics from the namespaced and un-namespaced
URLs.

Since the namespacing has been deployed this code is not necessary
anymore because the URLs without namespace are being redirected by the
router to their namespaced counterparts. This happens because we've
added redirect items for the topics to the content store from
collections-publisher.

Trello: https://trello.com/c/EoWPp4qZ/187-move-topics-to-topic