Extract content item finding to class
In all cases in this application, we wish to raise a HTTPNotFound error
when the requested content item doesn't exist. Using
`ContentItem.find!` we make sure we don't inadvertently use
`ContentStore#content_item` (without bang) in the future.

Also cleans up the code and allows for future validation and
transformation.