Fix subsection-to-section animation for curated pages
Curated pages have no A to Z heading, but the animation still expected
it to have a margin.

When showing a curated list, we animate to the gutter width (30px). The
solution chosen here is not very elegant. Some refactoring of the
animation code later might be a good idea, so we can remove duplication
between css and JS.