Handle base_path's in curated list data.
Currently the curated list data contains an array of contentapi API urls that
are used to match up against the data returned from contentapi or rummager. We
are pulling this data from Rummager for Topics, and will soon be doing the same
for mainstream browse pages. It therefore makes sense for the data in
content-store to use base_paths instead of api_urls.

This therefore updates collections to handle this data sent in either format so
that we can make the switch on the publisher end.