Refactor MainstreamBrowsePage and Topic
`MainstreamBrowsePage` and `Topic` have a lot in common. To avoid
introducing abstractions we chose earlier not to use inheritance for
these classes.

This PR instead delegates some of the data things to the `ContentItem`
class, which is passed around as a real object instead of a hash. It
cleans up some duplication while still being readable.