Remove browse page from breadcrumbs
The browse index page at `/browse` currently displays a breadcrumb as “Home > Browse”, thereby linking back to itself. For consistency with breadcrumbs elsewhere on GOV.UK, this commit removes this link, so that the breadcrumb on the index page is simply “Home”.

The decision was taken not to display the breadcrumb as "Home > Browse" when a user clicks on a top-level browse page, because direct visits to the browse index page are very low and most users end up on browse pages from the GOV.UK home page, by clicking a link to go directly to a top level browse page. Therefore, the breadcrumb only changes once a user clicks on a second level browse page link.