Properly create subscriberlists
These parameters tell email-alert-api to create a "subscriberlist" with
certain parameters. The parameters determine when the user receives an
email. This used to be done by saying specifying the URL of a topic the
user wants to subscribe to. Because URLs always change, this is being
moved to be based on `content_id` in the `links` hash.

This app was still sending both `links` and `tags`, when
email-alert-api only accepts one of these. Since a deploy this morning
email-alert-api has been suddenly rejecting the payloads (as if the
validation just got introduced) - we don't know why.