Lookup facet organisations by topic content_id
Rummager has new fields to allow us to filter the so called
`specialist_sectors` by `content_id`. The new field is called
`topic_content_ids`, which refers to the correct entity (Topics) and
also contains content IDs instead of slugs. This means that we will be
able to change the URL of a topic/subtopic without breaking search.

This commit also:
- makes sure we use content IDs instead of slugs in `ListSet`
- removes the unused parameter `topic_type` from
  `Topic::ChangedDocuments`.