Fix all Rubocop linter warnings
This is in preparation for migrating the build to Jenkins 2. Jenkins 2 will require
us to add new build scripts which could optionally include Rubocop linting, so this
seems like a convenient point to add linting to this project.

Most of the Rubocop errors were trivial and could be fixed automatically.

One fix used Ruby's keyword arguments with no defaults, which are only supported by
Ruby 2.1 and above, so the Rubocop language level has been raised to 2.3 to be
consistent with the language level of the project.