Fix flaky integration test
MainstreamBrowsingTest was failing intermittently (example seed: 424). It would
pass if it happened to be run after one of the other integration tests, but
would fail otherwise.

The problem was that the test was using the Rails 'get' helper to get the page,
rather than the Capybara 'visit' function. This meant that the Capybara 'page'
variable was only defined if another integration test had already run, so the
breadcrumb assertions were not testing the page they were intended to.