Handle link groups with no title
Currently, if the “services and information” controller encounters a link group without a title then it returns a 500 error. This usually happens when links are not expanded by rummager due to unsynced data between rummager and the publishing-api. This commit removes any such groups to prevent such issues from affecting the “services and information” pages.