Add rcov to Jenkins build
The old Jenkins build used to publish the rcov reports, but this was
missed when migrating to Jenkins 2. This commit configures rcov report
publishing as part of the pipeline build.

The default coverage metrics are used, which means that the build does
not fail if coverage is low, so rcov is really just being used to
generate a graph of coverage over time.