Create new navigation page for taxons
This commit adds the backend work to display the new navigation page for
taxons. It's missing styles to match the ones in the prototype, but
provides the data needed.

The new navigation pages will always be on in test and development
modes, but will need an environment variable set in other environments
in order to be displayed.