Ignore linting errors for now
We would like to break down the work on the new navigation pages so more
developers can help out. These styles will be completely refactored in
subsequent tasks, so we are only temporarily disabling some linting
errors in order to be able to merge these changes and continue the work.

This is a temporary commit.