Fetch breadcrumbs from the server in browse pages
We have recently enabled click tracking on breadcrumbs via data
attributes. These data attributes are added by `static` and are needed
to know where people click. The old functionality in `collections` meant
that we would be adding list items to the breadcrumbs ourselves in the
app and therefore not including any of the necessary data attributes.

This commit re-fetches the breadcrumbs from either cache or the server
rather than modifying it.

This way, we always render breadcrumbs that have the expected attributes
set.

The tests are now simpler as well, since the logic of fetching
components from the server is already tested.