Rename all taxon base paths in test helpers
Taxons' base paths will have from now on a prefix of the root taxon in
their taxonomy tree. This means that they will no longer be prefixed
with `alpha-taxonomy`, which was a temporary prefix that was not meant
to be public.

This commit renames existing taxon base paths in the test helpers.