Render alphabetical lists on taxon pages as ordered lists
Mark the topic and taxon content lists as `ol` rather than `ul` to mark
them as ordered (alphabetical) lists. This is more informative and can
be used by browsers like screen readers to provide more navigation
control.

This does not change the visual appearance of the lists in a standard
browser.