Track clicks on navigation page accordions in Google Analytics
Merge all the recent changes from the accordion code in
service-manual-frontend except for the BEM CSS selectors.

These changes include Google Analytics tracking for the accordion, which
we will use to investigate how users use the accordion in the education
navigation A/B test.

The long term plan is to extract the accordion to a common module, but
we will only do this once the final design of the taxonomy navigation is
complete.