Filter out content belonging to a collection
Document collections should be surfaced on navigation pages, but not the
content tagged to them.

We have seen from lab research that surfacing all the content makes the
list view harder to navigate.

Some document collections will be removed as they are now covered by
taxons, but other that group the same type of thing will remain.

This commit makes sure we filter out all content that already belongs to
a document collection.