Use content store instead of Rummager in topic and browse pages
Content is linked to parent topics and mainstream browse pages in the
publishing API and content store, so the content store can be used to
populate the subtopic and second-level browse pages rather than
Rummager.

This reduces the number of requests to Rummager from
collections, and is a step towards removing it from collections
entirely.