Track the type of taxon page being visited in Google Analytics
Add meta tag to taxon pages which identifies whether the page is a grid,
an accordion or a leaf node. This will help us analyse whether users
are able to find what they need quickly, or if they navigate in circles
because they are struggling to find content in the new navigation.