Track all nav pages as `finding` pages in Google Analytics
Add meta tag which marks taxon, browse, topic and services & information
pages as being in the `finding` (navigation) stage of a user journey, as
opposed to the `thing` (content) stage.

This will enable us to analyse how users navigate the site so that we
can work out whether the new navigation pages are helping users find
what they need.