Add Overview section with content to accordion
This commit makes sure the first section of the accordion view is an
Overview section with content tagged to the accordion taxon level.

When there is no content, we omit this section and only show child
taxons and their content.

The copy might change, but for now we should go ahead with this one.