Filter document collections and content items
This commit adds rules around filtering document collections and content
tagged to document collections.

The reason for these rules is that document collections are all
currently under review. Most of them should disappear and its content
solely tagged to taxons, but others will stay on. Some of the existing
document collections have dozens or hundreds of content items tagged to
them. If we decided to show those content items without any control, the
navigation pages would become unusable.

In other cases, the document collections are manageable, which means
it's fine to show the content tagged to them.

The rules are as follows:
- if a document collection is whitelisted (with `surface_collection:
  true`), then that document collection should show up;
- if a document collection is not supposed to be surfaced (with
  `surface_collection: false`), then it should not show up in the
  navigation pages;
- if documents tagged to a collection should be surfaced (with
  `surface_content: true`), then the content tagged to a document
  collection should show in the navigation pages;
- finally, if documents tagged to a collection should not be surfaced
  (with `surface_content: false`), then the content tagged to a document
  collection should not show in the navigation pages;

Finally, these rules also allow us to make sure there are no taxons
without content showing up. There are cases where the only content
tagged to a taxon is in a document collections, which wouldn't show. By
manually allowing to control this, we make sure that doesn't happen.