Enable ABTest-Benchmarking
With this new A/B test, we will be able to enable Mouseflow for all 60
users in the benchmarking round. This will allow us to collect data on
how the navigation pages are used and generate a heatmap in the end in
order to help us understand what areas of the site will be improved.