Add email alert buttons
Add template for email alert buttons. This still needs to be integrated
with the backend. I've used the styling used in government-frontend to
serve the travel publisher pages. I think it still makes sense to keep
the subscriptions section.

I've changed the visually hidden header to subscription rather than
subscriptions as currently we only have email alerts, although in the
future I think we plan to add in RSS feeds as well. I'm keeping the
subscriptions class, as I think at some point this will probably be
consolidated and don't want to make new classes.

I've added in a 3px margin to override the negative margins in the rest
of collections on the taxon pages. I don't think this is the cleanest
CSS, but touching the rest of the CSS in collections at the moment is
out of scope.