Allow for fragments for taxon redirects
This commit fixes an issue related to redirecting to a taxon page with a
fragment. Commit 7529768e00b75dfbaf80e2a92508c46f13785a96 added a
redirect with a fragment, however the resulting URL is:

```
/education/funding-and-finance-for-students%23/education/student-grants-bursaries-scholarships
```

By passing in an `anchor` param to `redirect_to`, we allow the redirect
to work and also make sure the user will be positioned in the expected
fragment.