Update accordion tracking label
When clicking an accordion heading, we send a tracking event. We
previously tracked where on the header we clicked:

- on the heading link
- on the +/- icon
- anywhere else (whitespace or description)

In analysis, this amount of information in the tracking label was too
verbose, so we are removing it.

Instead, we now return just the heading, and its position in the
accordion. For example, if we have a taxon entitled “Phonics” as the
fourth taxon in the accordion, if that section is expanded the tracking
label will be `4. Phonics`.