Test email alert link and refactor logic
This commit adds tests for the email subscription link on navigation
pages. It also refactors logic around when to show the email
subscription link on accordion pages so it's clearer. On accordion
pages, the "General Information and Guidance" section should not have
the email subscription link on it, so I've renamed the method to
indicate that.