Fix deprecation warnings on rendering
This fixes the following warnings:

```
.DEPRECATION WARNING: `render :text` is deprecated because it does not
actually render a `text/plain` response. Switch
to `render plain: 'plain text'` to render as `text/plain`, `render html:
'<strong>HTML</strong>'` to render as `text/ht
ml`, or `render body: 'raw'` to match the deprecated behavior and render
with the default Content-Type, which is `text/
html`.
```