Add description for News pages
Worldwide News pages will be tagged to World Loctaions and as such
will display on World Location navigation pages. We want a specific hard
coded summary of "Updates, news and events from the UK government in
[country]" to display under these.

This commit adds a description helper to display the summary.

The only way to determine if a page is a worldwide news page is to test
the base path returned by Rummager.