Set specific order for world taxonomy accordion content
For the new world taxon pages (eg /world/germany), we require the child taxons to be displayed in a specific order. This commit adds short-term solution. Ideally we would like to be determine the order somewhere else in the publishing platform, but given the time constraints, this might be the best solution at the moment.

We rely on the title of the taxons so that's quite brittle, but not sure if there's anything better?