Redirect `/world/all` to `/world`
As part of the worldwide taxonomy migration we currently need to leave
the `/world` page rendered by Whitehall as there isn't an equivalent, or
suitable format within Collections. We do need the `/world` taxon as
part of the taxonomy but it won't be rendered.

The solution for this is to publish the taxon to `/world/all` and
redirect `/world/all` to `/world` which will fall through the router to
Whitehall.

This commit adds a redirect to routes to accomplish this.