Track clicks on content items from subtopics
This commit adds the Google Analytics tracking module to subtopic pages
and adds a number of tracking data attributes to all content items'
links so that we can run reports to determine how these links are being
used.

Note that we track the position of the links relatively to the
subsection they are on. For example, the 3rd link on the 2nd subsection
will have a position of '2.3'.