Redirect Childcare and Parenting topics and browse
This commit adds redirects to all topics and browse pages that show
content related to Childcare and Parenting. The redirects will only
happen for users in the B variant of the EducationNavigation A/B test
and are in place to keep navigation consistent and prevent those users
from seeing old navigation patterns.

Trello: https://trello.com/c/tOQACFQQ/71-create-a-mapping-of-legacy-taxonomies-taxons-for-childcare-and-parenting-so-we-can-redirect-them