Don't report timeouts to Sentry (Part II)
In https://github.com/alphagov/collections/pull/383 I tried filtering
timeout exceptions, because they're mostly noise and distract from real
errors.

That PR doesn't work because Rails rescues and re-raises errors. This
means that the error we see is actually a
`ActionView::Template::Error`. We can't filter out those errors because
it's likely that any real problems would be masked by that.

The solution is to look at the `cause` of the exception. This is the
exception that's the original cause, see:

```rb
begin
  begin
    raise ArgumentError, "Hello!"
  rescue
    raise StandardError
  end
rescue => e
  puts e.class.name # "StandardError"
  puts e.cause.class.name # "ArgumentError"
end
```

I've proposed a solution on the Raven repo:
https://github.com/getsentry/raven-ruby/issues/765, but it's a bit of a
risky change to make in that gem.

This commit is a experiment to see if we can filter by looking at
`cause`. If it works we can push it up to the gem, and eventually to
the Sentry client.

https://trello.com/c/lqHmGMEB