Remove extra filtering code for exceptions
In https://github.com/alphagov/collections/pull/388 I fixed a bug and
introduced another: the `e` argument to the proc isn't an exception,
but a `Raven::Event`. To make sure we don't do this again, this commit
reverts the changes so that error reporting will definitely work. In a
follow up PR we'll try again, but this time with tests.

https://trello.com/c/h7Rt76jU