Add learn-to-drive-a-car task list page
This page is the first item in the public beta of task list pages run by the
modelling-services team.

The content item for this page is [manually published from collections-publisher](https://github.com/alphagov/collections-publisher/pull/250)

The json file recording the content is deep symbolized to facilitate working
with components which use symbolized keys.

This page is expected to replicate [the prototype](https://govuk-services-prototype.herokuapp.com/services/learn-to-drive-a-car)