Remove task list A/B test from /browse
The task list `Learn to drive a car: step by step` was manually added to the `/browse` page a couple of months ago through an A/B test.

As it has been live for a while now with no issues, we no longer need to display this using an A/B test.

We will get a content designer to add a proper link to this task list using the Collections Publisher instead. Once this is ready we will deploy our changes.