Update tasklist json with the new schema
The component schema has changed. We've removed nested groups to indicate steps
that can be progressed concurrently, and adding a `logic` parameter to indicate
whether a step should be done in addition to (and) or instead of (or) the
step that precedes it.

The content is the same apart from this.

This is still hardcoded content, but we are progressing publishing this at the
moment.