Only show email alert option if taxon is live
We are planning on publishing the taxonomy so that all the taxonomy is live, but in different phases, e.g: 'alpha', 'beta' and 'live'. This will make it easier for teams who need to work and prototype with the draft taxonomy, e.g: Topic Pages team. We want to make sure that pages for alpha and beta don't allow users to subscribe to email alerts.