Add phase check for email alerts
This commit reverts #516, which in turn reverted #505, #507 and #508.

In preparation for publishing the taxonomy in phases of alpha, beta or live, we want to make sure that we only show email alert buttons for taxons that are live and that if a taxon is not live we stop it being indexed by search engines.