Add integration tests for a live taxon
The design of the taxon pages is changing. We will no longer
be rendering different types of page (grid, accordion, list)
depending on the number of sub-topics a taxon has. Instead
all taxon pages will look the same.

This means that we can simplify the integration tests.

The test specifies which content supergroup to filter so that it's clear
that we are stubbing guidance_and_regulation results