Add tests for outstanding content purpose supergroups
Three supergroups require the published date and document type to be displayed
on topic pages: `news_and_communications`, `policy_and_engagement` and
`transparency`. This adds tests to check that the Taxon Presenter gets this
data for these content purpose supergroups and adds additional check to the
taxon browsing integration tests.