Never show non-taxon things as taxon
It seems like we have been showing the step by step pages with the
taxon layout recently. I think this is caused by the
`DocumentTypeRoutingConstraint` not finding the content item
temporarily, and thinking it's a taxon (the fallback in the routes).

By crashing we get 1) insight into the error, 2) Fastly will serve the
previous (correct) page from the mirror, instead of a wrong page.