Make highlight boxes take full width
Use mobile first CSS
When highlight boxes are full width on mobile, they should take up the full width of the space they're in. This commit removes the padding that was preventing them from doing this.
Highlight boxes should take up full width of the section they're placed in. Added a negative margin to counteract the padding-right on the last item