Display the closing dates for consultations
This gets the closing date for the consolation from content-store.
A possible enhancement is to add the closing date to the data
retrieved from rummager, but as there will no more than 5
consultations displayed per page, it didn't feel like a priority to
make this change right now.

A description of the date will be added in a future commit.