Clean up taxon-element-styles mixin
The taxon-element-styles mixin was created so it could be easily applied to different taxon pages.
Now that taxons are separate from worldwide pages, world pages no longer need those styles.
It therefore makes sense to remove the mixin and apply styles directly to the taxon grid.

This commit also removes some CSS which was designed to improve link touch area, but there was no supporting research to suggest this made an improvement.