Move see more link to common partial
The see more link is always tied together with the document list on topic pages.
It makes sense to move the see more link into the document list partial. This will also make it easier to do the work to adjust the news and communication section to the new layout with featured photo item.