Show See More under featured item if no more docs
If there is only one news and communication item for a taxon, this item will be displayed as the featured item.
We therefore need to check how many documents there are before displaying the see more link, so that the link is shown under the featured item unless there is a document list.