Render organisations with custom logos
If the logo details returned from rummager does not contain a
crest and a url (for custom logos) then that organisation should be displayed in a document list.

Similarly, if the logo details returned from rummager contain a custom logo url, for now we will also show
these in a document list with no logo visible. This is because otherwise we would need to resize the custom logos,
most of which contain text - this would make the text in the image inaccessible.