Remove top margin on the basic organisation list
There is a large gap between the last logo list item and the
beginning of the text only list in the organisations section.

This is caused by passing the "margin-top" attribute to the
document-list component in the govuk_publishhing_components gem.
"margin-top" sets the margin for the list as a whole and not
between individual list items.

Removing the margin-top attribute for organisations, fixes that
problem. 

Added a margin-bottom to the section-heading class in case there is a 
taxon that doesn't have any organisations with standard logos. In that
case the list item would have be rendered too close to the Organisation
section heading.