Add tracking to organisations without logos
We are treating the list of organisations as one list in the tracking, despite them being separate in the code (with and without logos).
This means we don't want the track_action number to reset for organisations without logos.
We want the number to continue on from the end of the organisations with logos.

Note: tracking for organisations with logos will be completed at a later date, as this will require a change to the component itself.