Separate out promoted content in guidance section
We need to be able to track different styles of content independently.
As we were passing everything via the `document_list` the different
styles were being counted as one list.

Separating out promoted content allows us to reset the list position of
the content for tracking purpose, so that the track-action number starts
at 1 again for plain list items.