Always show see more link for supergroup section
We always want to show the see more link for each supergroup section, even if there are only enough documents for the highlight boxes.
This removes the documents.any? check which was hiding the see more link if there were no documents for the document list.
We can remove this as the document list component has it's own check for items.any?

Also amended the taxon browsing test so the stubbing methods only returns 2 documents. Taxon browsing test checks whether the see more link is present, so this test will fail if we manage to introduce this bug again.