Stop using document list for non-ministerial
We need to stop using the document list for non-ministerial organisations (i.e: ones without logos).
The limitations of the component would make it more complicated to add the View All toggle
Also, the rest of the list (without the toggle) is just a link and some description text - this should be easy to style without a component

This commit removes all logic related to non-ministerial organisations and brings it together into one all_organisations method.
The view should still look the same as styling has been reapplied