Add organisation atom feeds
We provide similar atom feeds to those in Whitehall - but not identical.

The ID format has changed from the current one of the form:
  `tag:www.gov.uk,2005:DocumentCollection/231443` (which is coupled to whitehall
  internals)
to:
  `https://www.gov.uk/government/publications/universal-credit-understanding-its-impact-on-the-labour-market#2018-06-08T11:29:52+01:00`
  which will change when a document is updated or moved.  This is similar to
  that used by [travel advice feeds](https://github.com/alphagov/government-frontend/blob/9496ce14e3cbe117ea8ae9940504dd48e40dd5d1/app/views/content_items/travel_advice.atom.builder#L7).

This works with locales, however the only thing that changes with the language
is the feed information.  Entries in the feed are not locale sensitive (which
is actually the same behaviour as Whitehall at present.).

We can no longer provide long form feed content as Rummager doesn't let us
obtain that.  We fall back to providing a title and a summary.