Fix org filter not searching ministerial departments
The organisation index page filter broke when we moved the organisation logo component to the gem, because it needs to look for a specific class within that component.

Our JS tests didn't pick this up because we stub all the markup. This isn't ideal, but for now we can update the JS to look for the new updated class (which hopefully won't change).