Use smaller image sizes for people images
By default, the image URL that is in the content item is the largest possible size. This is far too big for the sizes we're showing on the organisation pages and means if there are a lot of people linked to an organisation, the images are slow to load.

We can instead use smaller versions of the images. I've upped the size slightly from what we use currently (s216_ to s465_) as they're a bit blurry on the current live pages.