Harden the code around consultation closing dates
The Content Store can redirect to another content item. When looking
up the consultation closing dates, if this happens, stop, as it means
that something's not quite right.

If this happens, the page will just omit the closing date.

This was spotted on Integration, as the Content Store was returning
the /government prefix route instead of a consultation, probably
because of issues with replicating data between environments.