Add Welsh translations to org pages
We were previously showing Welsh content items for translated content, but not
the bits that were part of the page framework.

I've brought over translations from cy.yml in Whitehall where appropriate and
set up the controller to switch locales if needed.

We should get cy.yml reviewed to fill in the gaps if we can, but this should be
an improvement.

When we make rummager requests, we shouldn't rely on the org slug being legit
as it might have a locale on the end - which we can trim off in this case.

NB Not every title in Whitehall org pages is translated, so we're not losing by
not being 100% covered. See "High profile groups" on https://www.gov.uk/government/organisations/department-for-work-pensions.cy